+++ IPFMDao/src/th/co/ais/ipfm/dao/hibernate/HibernateIPAccessListTNPDao.java
@@ -1,3 +1,4 @@
+ /* This application remediation was done for embedded Oracle SQL to make it compatible with PostgreSQL with Newt DMAP Version: v9.1.0.1_v8.4.2.9 on Date: 27-Jun-2025 */
 package th.co.ais.ipfm.dao.hibernate;
 
 import java.util.List;
@@ -21,13 +22,18 @@ public class HibernateIPAccessListTNPDao extends HibernateGenericDao<IpUrAccessL
 	@Override
 	public String getSubUrNo(String urNo) {
 		Session session = getSessionFactory().getCurrentSession();
+/**
+DMAP TAG: Query converted Needs Manual Intervention : Identifier169
+DMAP ConvertedQuery - SELECT coalesce(MAX(CAST(T1.SUB_UR_NO AS numeric)), 0) AS MAX_SUB_UR_NO FROM IP_UR_ACCESS_LIST_TNP T1 WHERE T1.UR_NO='urNo.trim()'
+**/
+
 		String subUrNo =  (String) session.createSQLQuery("SELECT NVL(MAX(T1.SUB_UR_NO),0) AS MAX_SUB_UR_NO FROM IP_UR_ACCESS_LIST_TNP T1 WHERE T1.UR_NO='"+urNo.trim()+"'")
 				.addScalar("MAX_SUB_UR_NO", Hibernate.STRING).uniqueResult();
 		return subUrNo;
 	}
 
 	@Override
-	public IpUrAccessListTnp getID(IpUrAccessListTnpId id){
+	public IpUrAccessListTnp getID(IpUrAccessListTnpId id){ // DMAP Comment : Dead Code Detected - The Following Method has no reference getID
 		Session session = getSessionFactory().getCurrentSession();
 		Criteria criteria = session.createCriteria(IpUrAccessListTnp.class);
 		criteria.add(Restrictions.eq("id", id));
@@ -36,7 +42,7 @@ public class HibernateIPAccessListTNPDao extends HibernateGenericDao<IpUrAccessL
 	}
 
 	@Override
-	public List<IpUrAccessListTnp> findByURNo(String urno) {
+	public List<IpUrAccessListTnp> findByURNo(String urno) { // DMAP Comment : Dead Code Detected - The Following Method has no reference findByURNo
 		Session session = getSessionFactory().getCurrentSession();
 		Criteria criteria = session.createCriteria(IpUrAccessListTnp.class);
 		criteria.add(Restrictions.eq("id.urNo", urno));
@@ -47,29 +53,44 @@ public class HibernateIPAccessListTNPDao extends HibernateGenericDao<IpUrAccessL
 		return criteria.list();
 	}
 	@Override
-	public void cancelImpact(String urNo) {
+	public void cancelImpact(String urNo) { // DMAP Comment : Dead Code Detected - The Following Method has no reference cancelImpact
 		Session session = getSessionFactory().getCurrentSession();
 		String sql = "UPDATE IP_UR_ACCESS_LIST_TNP set IS_IMPACT='N' WHERE UR_NO=?";
+/**
+DMAP TAG: Query converted but found same: Identifier168
+DMAP ConvertedQuery - UPDATE IP_UR_ACCESS_LIST_TNP set IS_IMPACT='N' WHERE UR_NO=?
+**/
+
 		session.createSQLQuery(sql).setString(0, urNo).executeUpdate();
 		
 	}
 	@Override
-	public void deleteByUrNo(String urNo) {
+	public void deleteByUrNo(String urNo) { // DMAP Comment : Dead Code Detected - The Following Method has no reference deleteByUrNo
 		Session session = getSessionFactory().getCurrentSession();
 		String sql = "DELETE FROM IP_UR_ACCESS_LIST_TNP AT WHERE AT.UR_NO=?";
+/**
+DMAP TAG: Query converted but found same: Identifier167
+DMAP ConvertedQuery - DELETE FROM IP_UR_ACCESS_LIST_TNP AT WHERE AT.UR_NO=?
+**/
+
 		session.createSQLQuery(sql).setString(0, urNo).executeUpdate();
 		
 	}
 
 	@Override
-	public void delete(String urNo, String subUrNo) {
+	public void delete(String urNo, String subUrNo) { // DMAP Comment : Dead Code Detected - The Following Method has no reference delete
 		Session session = getSessionFactory().getCurrentSession();
-		String sql = "DELETE FROM IP_UR_ACCESS_LIST_TNP FW WHERE FW.UR_NO=? AND FW.SUB_UR_NO IN ("+subUrNo+")";
+String sql = "DELETE FROM IP_UR_ACCESS_LIST_TNP FW WHERE FW.UR_NO=? AND FW.SUB_UR_NO IN (" + subUrNo + ")";//String sql = "DELETE FROM IP_UR_ACCESS_LIST_TNP FW WHERE FW.UR_NO=? AND FW.SUB_UR_NO IN ("+subUrNo+")";
+/**
+DMAP TAG: Query converted: Identifier166
+DMAP ConvertedQuery - DELETE FROM IP_UR_ACCESS_LIST_TNP FW WHERE FW.UR_NO=? AND FW.SUB_UR_NO IN (subUrNo)
+**/
+
 		session.createSQLQuery(sql).setString(0, urNo).executeUpdate();
 	}
 	
 	@Override
-	public void updateSubUrStatus(IpUrAccessListTnpId id, String status,IpUser user) {
+	public void updateSubUrStatus(IpUrAccessListTnpId id, String status,IpUser user) { // DMAP Comment : Dead Code Detected - The Following Method has no reference updateSubUrStatus
 		Session session = getSessionFactory().getCurrentSession();
 		StringBuffer sql = new StringBuffer();
 		sql.append("update IP_UR_ACCESS_LIST_TNP ");
@@ -77,6 +98,11 @@ public class HibernateIPAccessListTNPDao extends HibernateGenericDao<IpUrAccessL
 		sql.append(" , UPDATE_BY=? ");
 		sql.append(" , UPDATE_DATE=SYSDATE ");
 		sql.append("where (SUB_UR_NO=? and UR_NO=?)");
+/**
+DMAP TAG: Query converted Needs Manual Intervention : Identifier165
+DMAP ConvertedQuery - update IP_UR_ACCESS_LIST_TNP set SUB_UR_STATUS=? , UPDATE_BY=? , UPDATE_DATE=statement_timestamp() WHERE (SUB_UR_NO=? and UR_NO=?)
+**/
+
 		session.createSQLQuery(sql.toString())
 			.setString(0, status)
 			.setString(1, user.getUserId())
@@ -86,11 +112,16 @@ public class HibernateIPAccessListTNPDao extends HibernateGenericDao<IpUrAccessL
 	}
 	
 	@Override
-	public void commitSubURDelete(String urNo, String changeType) {
+	public void commitSubURDelete(String urNo, String changeType) { // DMAP Comment : Dead Code Detected - The Following Method has no reference commitSubURDelete
 		Session session = getSessionFactory().getCurrentSession();
 		StringBuffer deleteSQL = new StringBuffer();
 		deleteSQL.append("DELETE from IP_UR_ACCESS_LIST_TNP "); 
 		deleteSQL.append("WHERE (UR_NO=?) AND (CHANGE_TYPE=?) ");
+/**
+DMAP TAG: Query converted but found same: Identifier164
+DMAP ConvertedQuery - DELETE from IP_UR_ACCESS_LIST_TNP WHERE (UR_NO=?) AND (CHANGE_TYPE=?)
+**/
+
 		session.createSQLQuery(deleteSQL.toString())
 			.setString(0, urNo)
 			.setString(1, changeType)
@@ -99,12 +130,17 @@ public class HibernateIPAccessListTNPDao extends HibernateGenericDao<IpUrAccessL
 	}
 
 	@Override
-	public void commitSubURUpdate(String urNo, String changeType) {
+	public void commitSubURUpdate(String urNo, String changeType) { // DMAP Comment : Dead Code Detected - The Following Method has no reference commitSubURUpdate
 		Session session = getSessionFactory().getCurrentSession();
 		StringBuffer updateSQL = new StringBuffer();
 		updateSQL.append("UPDATE IP_UR_ACCESS_LIST_TNP "); 
 		updateSQL.append("SET CHANGE_TYPE='' ");
 		updateSQL.append("WHERE (UR_NO=?) AND (CHANGE_TYPE=?) "); 
+/**
+DMAP TAG: Query converted Needs Manual Intervention : Identifier163
+DMAP ConvertedQuery - UPDATE IP_UR_ACCESS_LIST_TNP SET CHANGE_TYPE = NULL WHERE (UR_NO=?) AND (CHANGE_TYPE=?)
+**/
+
 		session.createSQLQuery(updateSQL.toString())
 			.setString(0, urNo)
 			.setString(1, changeType)
@@ -114,7 +150,7 @@ public class HibernateIPAccessListTNPDao extends HibernateGenericDao<IpUrAccessL
 	}
 
 	@Override
-	public List<IpUrAccessListTnp> waitDeleteUr(String urNo) {
+	public List<IpUrAccessListTnp> waitDeleteUr(String urNo) { // DMAP Comment : Dead Code Detected - The Following Method has no reference waitDeleteUr
 		Session session = getSessionFactory().getCurrentSession();
 		Criteria criteria = session.createCriteria(IpUrAccessListTnp.class);
 		criteria.add(Restrictions.eq("id.urNo", urNo));
@@ -123,7 +159,7 @@ public class HibernateIPAccessListTNPDao extends HibernateGenericDao<IpUrAccessL
 	}
 
 	@Override
-	public IpUrAccessListTnp findIpUrAccessListTnp(String urNo,String subUrNo) {
+	public IpUrAccessListTnp findIpUrAccessListTnp(String urNo,String subUrNo) { // DMAP Comment : Dead Code Detected - The Following Method has no reference findIpUrAccessListTnp
 		Session session = getSessionFactory().getCurrentSession();
 		Criteria criteria = session.createCriteria(IpUrAccessListTnp.class);
 		criteria.add(Restrictions.eq("id.urNo", urNo));
